requirements.txt
absl-py==2.1.0
adlfs==2023.12.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
azure-core==1.32.0
azure-datalake-store==0.0.53
azure-identity==1.20.0
azure-storage-blob==12.24.1
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
class-registry==2.1.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.11
decorator==5.1.1
demoji==1.1.0
distro==1.9.0
EasyProcess==1.1
emails==0.6
entrypoint2==1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.17.0
filters==1.3.2
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
googlesearch-python==1.2.5
GQLAlchemy==1.6.0
grpcio==1.70.0
gTTS==2.5.4
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
iso8601==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
jax==0.5.2
jaxlib==0.5.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
logging==0.4.9.6
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.4.1
more-itertools==10.5.0
mouse==0.7.1
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
mss==10.0.0
multidict==6.1.0
namex==0.0.8
neo4j==4.4.12
nest-asyncio==1.6.0
networkx==2.8.8
ngrok==1.4.0
nltk==3.9.1
notify2==0.3.1
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.66.3
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
plyer==2.1.0
portalocker==2.10.1
premailer==3.10.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
psutil==5.9.8
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
pymgclient==1.3.1
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.2.1
pyperclip==1.9.0
pypiwin32==223
pypng==0.20220715.0
PyQRCode==1.2.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==1.0.1
pyseto==1.8.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytweening==1.2.0
pytz==2024.2
pywebio==1.8.3
pywin32==308
pyzmq==26.2.0
regex==2024.11.6
replit==4.1.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==1.3.1
requirement.txt==0.1.0
rich==13.9.4
scipy==1.14.1
secure-smtplib==0.1.1
selenium==4.27.1
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools-rust==1.10.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
textblob==0.19.0
tflearn==0.5.0
tiktoken==0.9.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
tweepy==4.14.0
twilio==9.4.3
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
wavio==0.0.9
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
win10toast==0.9
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
